Move Setup Test DB to Suite, introduce env variables for test db connectivity
