+++ test/integration_test/create_habit_integration_test.go
@@ -4,7 +4,6 @@ import (
 	"bytes"
 	"database/sql"
 	"encoding/json"
-	"fmt"
 	_ "github.com/lib/pq"
 	. "github.com/onsi/ginkgo"
 	. "github.com/onsi/gomega"
@@ -14,24 +13,6 @@ import (
 	"net/http"
 	"net/http/httptest"
 )
-
-func setupTestDatabase() (*sql.DB, error) {
-	db, err := sql.Open("postgres", "host=localhost port=5433 user=postgres password=mysecretpassword dbname=go-api-backend-db-test sslmode=disable")
-	if err != nil {
-		return nil, fmt.Errorf("Failed to connect to the test db: %v", err)
-	}
-
-	err = db.Ping()
-	if err != nil {
-		return nil, fmt.Errorf("Failed to ping the test db: %v", err)
-	}
-
-	return db, nil
-}
-
-func clearTestDatabase(db *sql.DB) {
-	_, _ = db.Exec("TRUNCATE TABLE habit")
-}
 
 var _ = Describe("CreateHabitHandler", func() {
 	var (
